+++ tests/key/key.js
@@ -1,11 +1,11 @@
 var TTYUtil = require("../../export")(require("../../build/Release/ttyu"));
-var ttyu = new TTYUtil();
 
 process.on("message", function(test) {
+    var ttyu = new TTYUtil();
     ttyu.start();
     var testa = function(e) {
         ttyu.off(TTYUtil.EVENT.KEY, testa);
-        ttyu.stop();
+        ttyu.destroy();
         process.send(e);
     };
     ttyu.on(TTYUtil.EVENT.KEY, testa);